+++ README.md
@@ -1,19 +1,26 @@
+<!-- @format -->
 
 # INSTAGRAM CLONE APP
 
 This is a small learning project exploring the integration of Firebase and React for creating a full-stack application
 
+## Demo
+
+## Created with CodeSandbox
+
+To view life here 👉 [Demo](instagram-clone-by-liuba.netlify.app)
+
+To view code here 👉 [Sandbox](https://codesandbox.io/s/github/liuba-larousse/react-firebase-instagram-clone/tree/main)
+
 ## Learnings:
 
-✅ Authentication with Firebase 
+✅ Authentication with Firebase
 
-✅ Images upload and use with Firebase Storage 
+✅ Images upload and use with Firebase Storage
 
-✅ Real-time database update with Firebase’s Firestore 
-
+✅ Real-time database update with Firebase’s Firestore
 
 ## Installation and Usage
-
 
 ```bash
 npm install
@@ -21,12 +28,11 @@ npm install
 npm start
 ```
 
-View locallly on 
+View locallly on
 
 ```bash
 http://localhost:3000
 ```
-
 
 ## Resources
 
@@ -36,13 +42,8 @@ http://localhost:3000
 
 [Material UI docs](https://material-ui.com/)
 
-
-
 ## Contributing
+
 Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
 
 Please make sure to update tests as appropriate.
-
-
-## Created with CodeSandbox
-To view click here 👉 [Sandbox](https://codesandbox.io/s/react-firebase-instagram-clone-s9kmw)
