+++ public/index.html
@@ -38,7 +38,7 @@
     <meta property="og:type" content="website" />
     <meta
       property="og:image"
-      content="https://firebasestorage.googleapis.com/v0/b/instagram-clone-app-ebf8a.appspot.com/o/logo%2Ffavicon.ico?alt=media&token=84faf795-8963-4ee3-a0dc-8f4be1355783"
+      content="https://firebasestorage.googleapis.com/v0/b/instagram-clone-app-ebf8a.appspot.com/o/logo%2Finstaog.png?alt=media&token=dbec4b15-a6e9-4043-af12-569d409a9962"
     />
     <meta
       property="og:description"
@@ -46,7 +46,7 @@
     />
     <meta
       property="og:url"
-      content="http://liuba-larousse.github.io/react-firebase-instagram-clone/"
+      content="https://instagram-clone-by-liuba.netlify.app/"
     />
 
     <title>React App</title>
