+++ src/components/Upload.js
@@ -10,8 +10,9 @@ import {
 import { PhotoCamera } from '@material-ui/icons';
 
 import { makeStyles } from '@material-ui/core/styles';
-import { storage, db } from '../firebase';
-import firebase from '@firebase/app';
+import { AuthContext } from '../Firebase/context';
+// import { storage, db } from '../Firebase/firebase';
+// import firebase from '@firebase/app';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -66,82 +67,89 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function Upload({ username }) {
+export default function Upload() {
   const classes = useStyles();
+  const {
+    handleChange,
+    setCaption,
+    handleUpload,
+    progress,
+    username,
+  } = useContext(AuthContext);
 
-  const [caption, setCaption] = useState('');
-  const [image, setImage] = useState('');
-  const [url, setUrl] = useState('');
-  const [progress, setProgress] = useState(0);
-  const [imagePreview, setImagePreview] = useState();
+  // const [caption, setCaption] = useState('');
+  // const [image, setImage] = useState('');
+  // const [url, setUrl] = useState('');
+  // const [progress, setProgress] = useState(0);
+  // const [imagePreview, setImagePreview] = useState();
 
-  useEffect(() => {
-    setImagePreview(imagePreview);
-  }, [imagePreview]);
+  // useEffect(() => {
+  //   setImagePreview(imagePreview);
+  // }, [imagePreview]);
 
-  function handlePreview(image) {
-    storage.ref(`image_previews/${image.name}`).put(image);
-    storage
-      .ref('image_previews')
-      .child(image.name)
-      .getDownloadURL()
-      .then((url) => setImagePreview(url));
-  }
+  // function handlePreview(image) {
+  //   storage.ref(`image_previews/${image.name}`).put(image);
+  //   storage
+  //     .ref('image_previews')
+  //     .child(image.name)
+  //     .getDownloadURL()
+  //     .then((url) => setImagePreview(url));
+  // }
 
-  const handleChange = (e) => {
-    if (e.target.files[0]) {
-      console.log('file:', e.target.files[0]);
-      setImage(e.target.files[0]);
-      handlePreview(e.target.files[0]);
-    }
-  };
+  // const handleChange = (e) => {
+  //   if (e.target.files[0]) {
+  //     console.log('file:', e.target.files[0]);
+  //     setImage(e.target.files[0]);
+  //     handlePreview(e.target.files[0]);
+  //   }
+  // };
 
-  function handleUpload() {
-    const uploadTask = storage
-      .ref(`images/${image.name}`)
-      .put(image);
-    console.log('image:', image);
-    console.log('image name:', image.name);
-    uploadTask.on(
-      'state_changed',
-      (snapshot) => {
-        //progress function
-        const progress = Math.round(
-          (snapshot.bytesTransferred /
-            snapshot.totalBytes) *
-            100
-        );
-        setProgress(progress);
-      },
-      (error) => {
-        //error function
-        console.log(error);
-        alert(error.message);
-      },
-      () => {
-        //complete function
-        storage
-          .ref('images')
-          .child(image.name)
-          .getDownloadURL()
-          .then((url) => {
-            //post image inside db
-            console.log('url:' + url);
-            db.collection('posts').add({
-              timestamp:
-                firebase.firestore.FieldValue.serverTimestamp(),
-              caption: caption,
-              imageUrl: url,
-              username: username,
-            });
-          });
-        setProgress(0);
-        setCaption('');
-        setImage(null);
-        setImagePreview(null);
-      }
-    );
-  }
+  // function handleUpload() {
+  //   const uploadTask = storage
+  //     .ref(`images/${image.name}`)
+  //     .put(image);
+  //   console.log('image:', image);
+  //   console.log('image name:', image.name);
+  //   uploadTask.on(
+  //     'state_changed',
+  //     (snapshot) => {
+  //       //progress function
+  //       const progress = Math.round(
+  //         (snapshot.bytesTransferred /
+  //           snapshot.totalBytes) *
+  //           100
+  //       );
+  //       setProgress(progress);
+  //     },
+  //     (error) => {
+  //       //error function
+  //       console.log(error);
+  //       alert(error.message);
+  //     },
+  //     () => {
+  //       //complete function
+  //       storage
+  //         .ref('images')
+  //         .child(image.name)
+  //         .getDownloadURL()
+  //         .then((url) => {
+  //           //post image inside db
+  //           console.log('url:' + url);
+  //           db.collection('posts').add({
+  //             timestamp:
+  //               firebase.firestore.FieldValue.serverTimestamp(),
+  //             caption: caption,
+  //             imageUrl: url,
+  //             username: username,
+  //           });
+  //         });
+  //       setProgress(0);
+  //       setCaption('');
+  //       setImage(null);
+  //       setImagePreview(null);
+  //     }
+  //   );
+  // }
 
   return (
     <>
